feat: monitor hooks
These ones are implemented a little more minimally, because it is hard
to define a good implementation knowing what we know now. In comparison,
with the connection hooks, there are 4 well-defined points all of which
exist with the same arguments, no more, no less. If we need to modify
the interface in the future, then we can simply do so.